Consolidate syscalls; Filter out procfs events
When inspecting the syscalls via the linux audit framework, it's
recommended to have as trimmed down rules as possible. osquery
originally installed individual syscall watches without any custom
filters for each syscall. That can lead to a lot of implicit filtering
within kauditd since it needs to match an event through upto 35 syscalls
individually.

Libaudit allows adding multiple syscalls to be added in the same audit
rule data. This can lead to matching all of the syscalls in a single
match within the kernel. For the record, an audit rule contains an array
that contains information about what syscalls it needs to match within a
bitmap. Which means, multiple syscalls can be matches efficiently using
bitmasks. In practice this leads to much better cpu consumption.

For the audit rules of the form
```
-s always,exit -S execve
-s always,exit -S execveat
```
The new rule will be
```
-s always,exit -S execve,execveat
```

For a linux system, procfs is a read only filesystem for the most part
with an exception of /proc/sys which contains files that can be used to
temporarily modify the system behaviour. There are processes that rely
on the procfs heavily to get performance counters and state of the
machine. All of this is captured by the audit client as "read"
operations in osquery and are discarded eventually because they can not
be modified.

To overcome this additional compute, all events on the /proc except
/proc/sys can be ignored. For that a rule is installed for matching
rules on the /proc/sys and then after that, all events for the /proc
will be dropped before the main rule to match and parse it.

```
-s always,exit -S <syscalls> -F dir=/proc/sys
-s never,exit -S all -F dir=/proc
-s always,exit -S <syscalls>
```

This also leads to practical performance benefits since a machine may
have prometheus exporter tools running which poll procfs or even "top"
for that matter.